requirements.txt
flask==3.0.3
torch==2.4.0
torchvision==0.19.0
gunicorn==22.0.0
pillow==10.4.0
numpy==1.26.4
ultralytics==8.0.0
opencv-python==4.8.1.78
pyyaml==6.0.1 